requirements.txt
playwright
pytest
python-dotenv
langchain
langchain-groq
